+++ libraries/helpers_rhel.rb
@@ -1,0 +1,6 @@
+module Mysql
+  module Helpers
+    module Rhel
+    end
+  end
+end