+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'cookbooks@getchef.com'
 license 'Apache 2.0'
 description 'Provides mysql_service and mysql_client resources'
 
-version '5.3.7'
+version '5.4.1'
 
 supports 'amazon'
 supports 'redhat'
@@ -16,5 +16,6 @@ supports 'ubuntu'
 supports 'smartos'
 supports 'omnios'
 supports 'freebsd'
+supports 'suse'
 
 depends 'yum-mysql-community'