+++ spec/unit/mysql_server/debian/jessie/custom_stepinto_spec.rb
@@ -42,9 +42,9 @@ GRANT ALL PRIVILEGES ON *.* TO 'root'@'1.2.3.4/5' IDENTIFIED BY 'YUNOSETPASSWORD
   context 'when using default parameters' do
     it 'creates mysql_service[debian_jessie_custom]' do
       expect(debian_jessie_custom_run).to create_mysql_service('debian_jessie_custom').with(
-        :version => '5.5',
-        :port => '3308',
-        :data_dir => '/data'
+        :parsed_version => '5.5',
+        :parsed_port => '3308',
+        :parsed_data_dir => '/data'
         )
     end