+++ spec/unit/mysql_server/debian/jessie/default_spec.rb
@@ -13,9 +13,9 @@ describe 'mysql_test_default::server on debian-jessie' do
   context 'when using default parameters' do
     it 'creates mysql_service[debian_jessie_default]' do
       expect(debian_jessie_default_run).to create_mysql_service('debian_jessie_default').with(
-        :version => '5.5',
-        :port => '3306',
-        :data_dir => '/var/lib/mysql'
+        :parsed_version => '5.5',
+        :parsed_port => '3306',
+        :parsed_data_dir => '/var/lib/mysql'
         )
     end
   end