+++ spec/unit/mysql_server/debian/jessie/default_stepinto_spec.rb
@@ -49,9 +49,9 @@ SET PASSWORD FOR 'root'@'127.0.0.1' = PASSWORD('ilikerandompasswords');"
   context 'when using default parameters' do
     it 'creates mysql_service[debian_jessie_default]' do
       expect(debian_jessie_default_run).to create_mysql_service('debian_jessie_default').with(
-        :version => '5.5',
-        :port => '3306',
-        :data_dir => '/var/lib/mysql'
+        :parsed_version => '5.5',
+        :parsed_port => '3306',
+        :parsed_data_dir => '/var/lib/mysql'
         )
     end