+++ spec/unit/mysql_server/omnios/151006/default_spec.rb
@@ -13,9 +13,9 @@ describe 'mysql_test_default::server on omnios-151006' do
   context 'when using default parameters' do
     it 'creates mysql_service[omnios_151006_default]' do
       expect(omnios_151006_default_run).to create_mysql_service('omnios_151006_default').with(
-        :version => '5.5',
-        :port => '3306',
-        :data_dir => '/var/lib/mysql'
+        :parsed_version => '5.5',
+        :parsed_port => '3306',
+        :parsed_data_dir => '/var/lib/mysql'
         )
     end
   end