+++ spec/unit/mysql_server/rhel/6/default_spec.rb
@@ -13,9 +13,9 @@ describe 'mysql_test_default::server on centos-6.4' do
   context 'when using default parameters' do
     it 'creates mysql_service[centos_6_4_default]' do
       expect(centos_6_4_default_run).to create_mysql_service('centos_6_4_default').with(
-        :version => '5.1',
-        :port => '3306',
-        :data_dir => '/var/lib/mysql'
+        :parsed_version => '5.1',
+        :parsed_port => '3306',
+        :parsed_data_dir => '/var/lib/mysql'
         )
     end
   end