+++ spec/unit/mysql_server/suse/11.3/default_spec.rb
@@ -13,9 +13,9 @@ describe 'mysql_test_default::server on suse-11.3' do
   context 'when using default parameters' do
     it 'creates mysql_service[suse_11_3_default]' do
       expect(suse_151006_default_run).to create_mysql_service('suse_11_3_default').with(
-        :version => '5.5',
-        :port => '3306',
-        :data_dir => '/var/lib/mysql'
+        :parsed_version => '5.5',
+        :parsed_port => '3306',
+        :parsed_data_dir => '/var/lib/mysql'
         )
     end
   end